stats-command

+++ tracer/include/util.h
@@ -1,7 +1,7 @@
 #ifndef UTIL_H
 #define UTIL_H
 
-float calc_mili(struct timeval begin, struct timeval end);
+long calc_mili(struct timeval begin, struct timeval end);
 
 int count_char(char* string, char x);
