+++ tracer/src/util.c
@@ -3,12 +3,12 @@
 
 #include "../include/util.h"
 
-float calc_mili(struct timeval begin, struct timeval end)
+long calc_mili(struct timeval begin, struct timeval end)
 {
-    float seg = (float) end.tv_sec - (float) begin.tv_sec;
+    long seg = end.tv_sec - begin.tv_sec;
     seg *= 1000;
 
-    float micro = (float) end.tv_usec - (float) begin.tv_usec;
+    long micro = end.tv_usec - begin.tv_usec;
     micro /= 1000;
 
     return seg + micro;
